# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "Documentation of API endpoints for the application"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur Local"
        }
    ],"components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        
        }
    },
    "security": [
        {
            "BearerAuth": []
    
            }
        ],
    "paths": {
...
        "/api/v1/referentiels": {
            "get": {
                "summary": "Get all active referentiels",
                "description": "This endpoint is used to retrieve a list of active referentiels.",
                "operationId": "getActiveReferentiels",
                "tags": [
                    "Referentiels"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            },
                                            "etat": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/promotion": {
            "post": {
                "summary": "Create a new promotion",
                "description": "This endpoint is used to create a new promotion.",
                "operationId": "createPromotion",
                "tags": [
                    "Promotions"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "referentiels": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "date_debut": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "date_fin": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "referentiels": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/promotions": {
            "get": {
                "summary": "Get all promotions",
                "description": "This endpoint is used to retrieve a list of all promotions.",
                "operationId": "getAllPromotions",
                "tags": [
                    "Promotions"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "libelle": {
                                                "type": "string"
                                            },
                                            "date_debut": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "date_fin": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "referentiels": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/promotions/{id}": {
            "patch": {
                "summary": "Update a promotion",
                "description": "This endpoint is used to update a promotion's information.",
                "operationId": "updatePromotion",
                "tags": [
                    "Promotions"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Promotion ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "referentiels": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "date_debut": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "date_fin": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "referentiels": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/promotions/{id}/referentiels": {
            "patch": {
                "summary": "Update referentiels of a promotion",
                "description": "This endpoint is used to update the referentiels of a promotion.",
                "operationId": "updatePromotionReferentiels",
                "tags": [
                    "Promotions"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Promotion ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "referentiels": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "date_debut": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "date_fin": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "referentiels": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/promotions/current": {
            "get": {
                "summary": "Get the current promotion",
                "description": "This endpoint is used to retrieve the current promotion.",
                "operationId": "getCurrentPromotion",
                "tags": [
                    "Promotions"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "libelle": {
                                            "type": "string"
                                        },
                                        "date_debut": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "date_fin": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "referentiels": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/promotions/{id}/close": {
    "get": {
        "summary": "Close a promotion",
        "description": "This endpoint is used to close a specific promotion by its ID.",
        "operationId": "closePromotion",
        "tags": [
            "Promotions"
        ],
        "parameters": [
            {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                },
                "description": "The ID of the promotion to close."
            }
        ],
        "responses": {
            "200": {
                "description": "Promotion closed successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Promotion closed successfully."
                                }
                            }
                        }
                    }
                }
            },
            "404": {
                "description": "Promotion not found"
            },
            "500": {
                "description": "Internal Server Error"
            }
        }
    }
},
...
  }
}